v1.0.2
Improved cone tracing to properly use all 6 voxel grids. Earlier it used to go up a level only if it couldn't find an intersection on a lower level first. Now, it goes up a level after a fixed number of iterations and then up 4 more levels till it finds a hit.